tests will pass
If the script is in the top level directory, or is a command instead of a script, check the $(pwd)

Note - if a file exists in the top level folder with the same name as the command, this will explode. E.g. "echo something" and a file named "echo" will cause us to check for a shebang in the file "echo"