Update path match to be case sensitive (#480)
* Update path match to be case sensitive

e.g. https://batch.hail.populationgenomics.org.au/batches/163811/jobs/1

Analysis-runner was able to start this job, despite the provided script name `plot_alt_af.py` not being a case-specific match to the real name `plot_alt_AF.py`
The cause is (likely) the lack of case recognition in ZSH, which results in case insensitive matching being done by `os.path.exists`

* please the linter

* Bump version: 2.32.3 → 2.32.4

* corrected the syntax

* tests will pass

If the script is in the top level directory, or is a command instead of a script, check the $(pwd)

Note - if a file exists in the top level folder with the same name as the command, this will explode. E.g. "echo something" and a file named "echo" will cause us to check for a shebang in the file "echo"

* ARGH