change how AR webserver serves files to improve performance
previously the webserver was attempting to stream the response using
stream_with_context, but this method is intended for smallish text files
to be returned line by line. This made the webserver extremely slow to
respond with even medium size files, taking nearly a minute to serve a
32mb .wasm file.

The intent of streaming the response was to allow serving of large files
by not writing them to disk or storing in memory before returning them,
but functionally this would not have worked as it would take so long to
serve the file so as to be useless.

This change uses a more traditional method of returning a file using
flask's `send_file` method which will read the file into memory before
responding with it. This won't allow larger-than-memory files to be
served but functionally we've never needed that option anyway.