cache expensive network calls to improve performance
The webserver request handler had a couple of calls, one to load the
server config secret, and one to get the membership cache that were
relatively expensive, taking between 1 and 2 seconds each. This update
caches those calls for a short period of time so that web pages that
make many requests for small files will not have to make these expensive
network calls on each request.