SET-502 Improve performance of analysis runner web server when loading large files (#729)
* change how AR webserver serves files to improve performance

previously the webserver was attempting to stream the response using
stream_with_context, but this method is intended for smallish text files
to be returned line by line. This made the webserver extremely slow to
respond with even medium size files, taking nearly a minute to serve a
32mb .wasm file.

The intent of streaming the response was to allow serving of large files
by not writing them to disk or storing in memory before returning them,
but functionally this would not have worked as it would take so long to
serve the file so as to be useless.

This change uses a more traditional method of returning a file using
flask's `send_file` method which will read the file into memory before
responding with it. This won't allow larger-than-memory files to be
served but functionally we've never needed that option anyway.

* cache expensive network calls to improve performance

The webserver request handler had a couple of calls, one to load the
server config secret, and one to get the membership cache that were
relatively expensive, taking between 1 and 2 seconds each. This update
caches those calls for a short period of time so that web pages that
make many requests for small files will not have to make these expensive
network calls on each request.

* update gunicorn workers and threads settings

The cloud run for the analyis runner web server has been updated to 2
CPUs, according to gunicorn docs the values for threads and workers
should be between 2-4 x $(NUM CORES)

@see https://docs.gunicorn.org/en/stable/settings.html