change back to a streaming method for web server responses
PR #729 aimed to improve performance by removing the use of
stream_with_context from the response handling. While this did improve
performance it meant that we're now hitting the 32mb response size limit
that cloud run imposes for non-streaming responses.

This update reintroduces a streaming response, but using a different
method that doesn't have the same small chunk size that was causing the
performance issue previously. The docs on this are pretty minimal but
this is pulling together some info from both the flask docs on
streaming:
https://flask.palletsprojects.com/en/stable/patterns/streaming/

and the fast api docs on streaming a file-like binary response:
https://fastapi.tiangolo.com/advanced/custom-response/#using-streamingresponse-with-file-like-objects